Compiler warning fix for https://github.com/rism-digital/verovio/pull/1659#issuecomment-845235993

+++ src/MuseRecordBasic.cpp
@@ -505,7 +505,7 @@ void MuseRecordBasic::append(const char* format, ...) {
 
 	HumNum rn;
 
-	int len = strlen(format);
+	int len = (int)strlen(format);
 	for (i=0; i<len; i++) {
 		switch (format[i]) {   // Type to expect.
 			case 'i':
